- Slight change to current "GUI.java" file - Major changes to come either 03/18/2016 or 03/19/2016
Upcoming changes (03/18/2016 or 03/19/2016) will include the following: 
- "SearchView.java" (this source code only manages graphics)
- "SearchController.java" (this source code updates model & view) 
- "SearchModel.java" (this source code stores state for the model)
- "ReservationView.java" (this source code only manages graphics)
- "ReservationController.java" (this source code updates model & view)
- "ReservationModel.java" (this source code stores state for the model)

Note:
All of the upcoming source code has been written in another separate
project on my local computer. The only changes that I need to make are
to incorporate other elements such as the entity that retrieves flights
from the database (into the controller code so that it can be called
upon) and to revise the graphical design of the ReservationView.
Additionally, I also need to review the current code base of our project
to see how flights should be displayed in String form.